refactor(app): extract GuideGraph component
